requirements.txt
 langchain 
 langgraph 
 openai 
 python-dotenv